refactor: moved add-list-items.js to to-do folder, fix(addListItem): input parameter to accept value
