+++ blocks/to-do/to-do.js
@@ -1,5 +1,5 @@
-import handleKeyPress from '../../scripts/key-press.js';
-import addListItem from '../../scripts/add-list-items.js';
+import handleEnterKeyPress from '../../scripts/enter-key-press.js';
+import addListItem from './add-list-items.js';
 
 // Selectors
 const toDoAddDiv = document.querySelector('.to-do > div');
@@ -16,14 +16,15 @@ export default function decorate(block) {
   toDoAddDiv.appendChild(input);
   block.appendChild(orderedList);
 
-  // Specify addListItem function
+  // Specify addListItem function parameters
   const addToDo = (event) => {
-    addListItem(event, input, orderedList);
+    addListItem(event, input.value.trim(), orderedList);
+    input.value = '';
   };
 
   // Event listeners
   addBtn.addEventListener('click', addToDo);
   input.addEventListener('keypress', (event) => {
-    handleKeyPress(event, addToDo, input);
+    handleEnterKeyPress(event, addToDo);
   });
 }
