+++ scripts/enter-key-press.js
@@ -1,0 +1,6 @@
+export default function handleEnterKeyPress(event, func) {
+  // Check if the Enter key was pressed
+  if (event.key === 'Enter') {
+    func(event);
+  }
+}
